requirements.txt
# Core Dependencies
appium-python-client==2.11.5
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0

# MCP Integration
mcp==0.1.0

# AI/NLP
openai==1.10.0
anthropic==0.18.0

# Configuration
python-dotenv==1.0.0
PyYAML==6.0.1

# HTTP Client
aiohttp==3.9.1
httpx==0.26.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-html==4.1.1
pytest-mock==3.12.0

# Utilities
selenium==4.16.0
Pillow==10.2.0
lxml==5.1.0

# Type Hints
types-PyYAML==6.0.12.12